Fixing css and formating front code
