+++ docker-compose.yml
@@ -1,7 +1,4 @@
-version: '3.8'
-
 services:
-  # Backend API with Socket.IO
   backend:
     build: ./back
     ports:
@@ -19,7 +16,6 @@ services:
       retries: 5
       start_period: 30s
 
-  # Frontend React App
   frontend:
     build: ./front
     ports:
