+++ front/src/App.css
@@ -23,32 +23,74 @@ body {
   min-height: 100vh;
   padding: 2rem;
   text-align: center;
+  position: relative;
+}
+
+.home-container::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: 
+    radial-gradient(circle at 20% 80%, rgba(120, 119, 198, 0.3) 0%, transparent 50%),
+    radial-gradient(circle at 80% 20%, rgba(255, 107, 107, 0.3) 0%, transparent 50%);
+  pointer-events: none;
 }
 
 .hero-section {
-  background: rgba(255, 255, 255, 0.95);
-  padding: 3rem;
-  border-radius: 20px;
-  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.1);
+  background: rgba(255, 255, 255, 0.98);
+  backdrop-filter: blur(10px);
+  padding: 3.5rem;
+  border-radius: 24px;
+  box-shadow: 
+    0 25px 50px rgba(0, 0, 0, 0.15),
+    0 0 0 1px rgba(255, 255, 255, 0.2);
   margin-bottom: 3rem;
   max-width: 600px;
   width: 100%;
+  position: relative;
+  z-index: 1;
+  transition: transform 0.3s ease;
+}
+
+.hero-section:hover {
+  transform: translateY(-5px);
 }
 
 .hero-section h1 {
-  font-size: 3rem;
+  font-size: 3.2rem;
   color: #333;
   margin-bottom: 1rem;
   background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
   -webkit-background-clip: text;
   -webkit-text-fill-color: transparent;
   background-clip: text;
+  font-weight: 800;
+  letter-spacing: -0.02em;
+  position: relative;
+}
+
+.hero-section h1::after {
+  content: '';
+  position: absolute;
+  bottom: -8px;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 60px;
+  height: 3px;
+  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+  border-radius: 2px;
+  opacity: 0.7;
 }
 
 .hero-section p {
-  font-size: 1.2rem;
-  color: #666;
-  margin-bottom: 2rem;
+  font-size: 1.3rem;
+  color: #555;
+  margin-bottom: 2.5rem;
+  font-weight: 300;
+  line-height: 1.6;
 }
 
 .button-group {
@@ -59,14 +101,32 @@ body {
 }
 
 .nav-button {
-  padding: 1rem 2rem;
+  padding: 1.2rem 2.5rem;
   font-size: 1.1rem;
-  font-weight: bold;
+  font-weight: 600;
   border: none;
-  border-radius: 50px;
+  border-radius: 60px;
   cursor: pointer;
-  transition: all 0.3s ease;
-  min-width: 200px;
+  transition: all 0.4s cubic-bezier(0.4, 0, 0.2, 1);
+  min-width: 220px;
+  position: relative;
+  overflow: hidden;
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
+}
+
+.nav-button::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: -100%;
+  width: 100%;
+  height: 100%;
+  background: linear-gradient(90deg, transparent, rgba(255, 255, 255, 0.3), transparent);
+  transition: left 0.6s;
+}
+
+.nav-button:hover::before {
+  left: 100%;
 }
 
 .chat-button {
@@ -75,8 +135,8 @@ body {
 }
 
 .chat-button:hover {
-  transform: translateY(-3px);
-  box-shadow: 0 10px 20px rgba(255, 107, 107, 0.3);
+  transform: translateY(-4px);
+  box-shadow: 0 15px 35px rgba(255, 107, 107, 0.4);
 }
 
 .data-button {
@@ -85,8 +145,8 @@ body {
 }
 
 .data-button:hover {
-  transform: translateY(-3px);
-  box-shadow: 0 10px 20px rgba(84, 160, 255, 0.3);
+  transform: translateY(-4px);
+  box-shadow: 0 15px 35px rgba(84, 160, 255, 0.4);
 }
 
 .features {
@@ -98,21 +158,46 @@ body {
 }
 
 .feature-card {
-  background: rgba(255, 255, 255, 0.9);
-  padding: 2rem;
-  border-radius: 15px;
-  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.1);
-  transition: transform 0.3s ease;
+  background: rgba(255, 255, 255, 0.95);
+  backdrop-filter: blur(10px);
+  padding: 2.5rem;
+  border-radius: 20px;
+  box-shadow: 
+    0 15px 35px rgba(0, 0, 0, 0.08),
+    0 0 0 1px rgba(255, 255, 255, 0.3);
+  transition: all 0.4s cubic-bezier(0.4, 0, 0.2, 1);
+  position: relative;
+  overflow: hidden;
+}
+
+.feature-card::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  height: 3px;
+  background: linear-gradient(90deg, #667eea, #764ba2, #ff6b6b);
+  opacity: 0;
+  transition: opacity 0.3s ease;
 }
 
 .feature-card:hover {
-  transform: translateY(-5px);
+  transform: translateY(-8px);
+  box-shadow: 
+    0 25px 50px rgba(0, 0, 0, 0.15),
+    0 0 0 1px rgba(255, 255, 255, 0.3);
+}
+
+.feature-card:hover::before {
+  opacity: 1;
 }
 
 .feature-card h3 {
   color: #333;
-  margin-bottom: 1rem;
-  font-size: 1.3rem;
+  margin-bottom: 1.2rem;
+  font-size: 1.4rem;
+  font-weight: 700;
 }
 
 .feature-card p {
@@ -129,30 +214,37 @@ body {
 .page-header {
   display: flex;
   align-items: center;
-  gap: 1rem;
+  justify-content: space-between;
   margin-bottom: 2rem;
+  flex-wrap: wrap;
+  gap: 1rem;
 }
 
 .back-button {
-  padding: 0.75rem 1.5rem;
-  background: rgba(255, 255, 255, 0.2);
+  padding: 0.8rem 1.8rem;
+  background: rgba(255, 255, 255, 0.15);
+  backdrop-filter: blur(10px);
   color: white;
-  border: 2px solid rgba(255, 255, 255, 0.3);
-  border-radius: 25px;
+  border: 2px solid rgba(255, 255, 255, 0.2);
+  border-radius: 30px;
   cursor: pointer;
-  font-weight: bold;
-  transition: all 0.3s ease;
+  font-weight: 600;
+  transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+  font-size: 0.95rem;
 }
 
 .back-button:hover {
-  background: rgba(255, 255, 255, 0.3);
-  transform: translateX(-5px);
+  background: rgba(255, 255, 255, 0.25);
+  transform: translateX(-3px);
+  border-color: rgba(255, 255, 255, 0.4);
 }
 
 .page-header h2 {
   color: white;
-  font-size: 2rem;
+  font-size: 2.2rem;
   margin: 0;
+  font-weight: 700;
+  text-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
 }
 
 .content-box {
@@ -180,85 +272,6 @@ body {
   color: #666;
   margin-bottom: 2rem;
   font-size: 1.1rem;
-}
-
-/* Mock Chat Styles */
-.mock-chat {
-  background: #f8f9fa;
-  border-radius: 15px;
-  padding: 1.5rem;
-  margin: 2rem 0;
-  text-align: left;
-  max-width: 500px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.mock-message {
-  padding: 0.75rem;
-  margin-bottom: 0.5rem;
-  background: white;
-  border-radius: 10px;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-}
-
-.mock-message strong {
-  color: #667eea;
-}
-
-.mock-input {
-  display: flex;
-  gap: 0.5rem;
-  margin-top: 1rem;
-}
-
-.mock-input input {
-  flex: 1;
-  padding: 0.75rem;
-  border: 2px solid #ddd;
-  border-radius: 25px;
-  outline: none;
-}
-
-.mock-input button {
-  padding: 0.75rem 1.5rem;
-  background: #ccc;
-  color: #666;
-  border: none;
-  border-radius: 25px;
-  cursor: not-allowed;
-}
-
-/* Mock Data Styles */
-.mock-data {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  gap: 1.5rem;
-  margin-top: 2rem;
-}
-
-.data-card {
-  background: linear-gradient(135deg, #f8f9fa 0%, #e9ecef 100%);
-  padding: 1.5rem;
-  border-radius: 15px;
-  border: 2px solid #dee2e6;
-  transition: all 0.3s ease;
-}
-
-.data-card:hover {
-  transform: translateY(-3px);
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.1);
-}
-
-.data-card h4 {
-  color: #333;
-  margin-bottom: 0.5rem;
-  font-size: 1.2rem;
-}
-
-.data-card p {
-  color: #666;
-  margin: 0;
 }
 
 /* Responsive Design */
@@ -289,9 +302,5 @@ body {
     flex-direction: column;
     align-items: flex-start;
     gap: 1rem;
-  }
-  
-  .mock-data {
-    grid-template-columns: 1fr;
   }
 }
